ggnewscale-package {ggnewscale}	R Documentation

ggnewscale: Multiple Fill and Colour Scales in 'ggplot2'

Description

Use multiple fill and colour scales in 'ggplot2'.

Usage

To have more than one colour or fill scale in a plot using ggplot2, you add a new scale slot with new_scale and the following geoms will use a different scale than the previous ones.

Author(s)

Maintainer: Elio Campitelli eliocampitelli@gmail.com (ORCID)

See Also

Useful links:

https://eliocamp.github.io/ggnewscale/

https://github.com/eliocamp/ggnewscale

Report bugs at https://github.com/eliocamp/ggnewscale/issues

[Package ggnewscale version 0.5.2 Index]