ggnewscale: Multiple Fill and Colour Scales in 'ggplot2'

Use multiple fill and colour scales in 'ggplot2'.

Usage

To have more than one colour or fill scale in a plot using ggplot2, you add a new scale slot with new_scale and the following geoms will use a different scale than the previous ones.

See also

Author

Maintainer: Elio Campitelli eliocampitelli@gmail.com (ORCID)